+++ server/api/alist/v2.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 
 	"github.com/freshteapot/learnalist-api/server/api/i18n"
-	"github.com/freshteapot/learnalist-api/server/api/utils"
 	aclKeys "github.com/freshteapot/learnalist-api/server/pkg/acl/keys"
+	"github.com/freshteapot/learnalist-api/server/pkg/utils"
 	"github.com/gookit/validate"
 )
