+++ server/e2e/client.go
@@ -13,8 +13,8 @@ import (
 	"time"
 
 	"github.com/freshteapot/learnalist-api/server/api/alist"
-	"github.com/freshteapot/learnalist-api/server/api/utils"
 	"github.com/freshteapot/learnalist-api/server/pkg/api"
+	"github.com/freshteapot/learnalist-api/server/pkg/utils"
 )
 
 // Generated by curl-to-Go: https://mholt.github.io/curl-to-go
