+++ server/pkg/event/doc.go
@@ -2,8 +2,8 @@ package event
 
 import (
 	"github.com/freshteapot/learnalist-api/server/api/alist"
-	"github.com/freshteapot/learnalist-api/server/api/utils"
 	"github.com/freshteapot/learnalist-api/server/pkg/openapi"
+	"github.com/freshteapot/learnalist-api/server/pkg/utils"
 	"github.com/nats-io/stan.go"
 )
