+++ server/pkg/utils/error.go
@@ -1,0 +1,7 @@
+package utils
+
+import "errors"
+
+var (
+	ErrNotFound = errors.New("not.found")
+)
